# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Sofware Engineering Project Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/order": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Get all order",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.Order"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Create order",
                "parameters": [
                    {
                        "description": "Create Order Payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payload.CreateOrderPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Response"
                        }
                    }
                }
            }
        },
        "/api/order/self": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get self order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Get self order",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.Order"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/order/self/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get self order detail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Get self order detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.Order"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.AccountType": {
            "type": "string",
            "enum": [
                "ADMIN",
                "USER"
            ],
            "x-enum-varnames": [
                "AdminAccount",
                "UserAccount"
            ]
        },
...
        "domain.Order": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "discount": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "orderProduct": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.OrderProduct"
                    }
                },
                "orderProductData": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.OrderProductResponse"
                    }
                },
                "productPrice": {
                    "type": "number"
                },
                "shippingPrice": {
                    "type": "number"
                },
                "status": {
                    "$ref": "#/definitions/domain.OrderStatusType"
                },
                "totalPrice": {
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/domain.User"
                },
                "userId": {
                    "type": "string"
                },
                "voucherCode": {
                    "type": "string"
                }
            }
        },
        "domain.OrderProduct": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "orderId": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "productSizeId": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "sellerId": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.OrderProductStatus"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "domain.OrderProductResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "orderId": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "productDescription": {
                    "type": "string"
                },
                "productName": {
                    "type": "string"
                },
                "productOptionImageUrl": {
                    "type": "string"
                },
                "productOptionLabel": {
                    "type": "string"
                },
                "productPrice": {
                    "type": "number"
                },
                "productSize": {
                    "type": "string"
                },
                "productSizeId": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "sellerId": {
                    "type": "string"
                },
                "sellerLogoUrl": {
                    "type": "string"
                },
                "sellerName": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.OrderProductStatus"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "domain.OrderProductStatus": {
            "type": "string",
            "enum": [
                "PENDING",
                "DELIVERY_ON_THE_WAY"
            ],
            "x-enum-varnames": [
                "ORDER_PRODUCT_PENDING",
                "ORDER_PRODUCT_DELIVERY_ON_THE_WAY"
            ]
        },
        "domain.OrderStatusType": {
            "type": "string",
            "enum": [
                "PENDING",
                "PAYMENT_SUCCESS",
                "REFUND",
                "END",
                "CANCEL"
            ],
            "x-enum-varnames": [
                "ORDER_PENDING",
                "ORDER_PAYMENT_SUCCESS",
                "ORDER_REFUND",
                "ORDER_END",
                "ORDER_CANCEL"
            ]
        },
...
        "domain.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "domain.User": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "google_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "profileUrl": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/domain.AccountType"
                },
                "updateAt": {
                    "type": "string"
                }
            }
        },
...
        "payload.CreateOrderPayload": {
            "type": "object",
            "required": [
                "addressId",
                "products",
                "shippingPrice"
            ],
            "properties": {
                "addressId": {
                    "type": "string"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/payload.OrderProductPayload"
                    }
                },
                "shippingPrice": {
                    "type": "number"
                },
                "voucherId": {
                    "type": "string"
                }
            }
        },
...
        "payload.OrderProductPayload": {
            "type": "object",
            "required": [
                "productSizeId",
                "quantity"
            ],
            "properties": {
                "productSizeId": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}